Update Home Path
I updated the home path (path redirected to when the My Photo Blog in the logo position) to redirect to a path "/view". The reason i changed it is so that i can use the "/" path to serve my files, therfore, when clicking any other path above that, the files will be automatically served.