Add functionality to pictures added.
I added a functionality to check the number of pictures added at the "add pictures" page before moving to the home page. The number of pictures will therefore determine the number of records to get from the database when routing back to the homepage. This ensures that a single picture is not gotten and displayed twice on the home page.