Update Database README.md
I updated the database information README file after decided that i needed to add an auto-incrementing id to the columns in my db table. I decided to add the id to use it to get only newly added pictures (on the home page) from the database while still working in a particular session.